+++ src/components/ConfirmDialog.react.js
@@ -58,6 +58,11 @@ ConfirmDialog.defaultProps = {
 };
 
 ConfirmDialog.propTypes = {
+    /**
+     * The ID of this component, used to identify dash components
+     * in callbacks. The ID needs to be unique across all of the
+     * components in an app.
+     */
     id: PropTypes.string,
 
     /**
@@ -84,7 +89,6 @@ ConfirmDialog.propTypes = {
      *  Set to true to send the ConfirmDialog.
      */
     displayed: PropTypes.bool,
-    key: PropTypes.string,
 
     /**
      * Dash-assigned callback that gets fired when the value changes.