+++ src/components/Dropdown.react.js
@@ -108,10 +108,16 @@ export default class Dropdown extends Component {
 }
 
 Dropdown.propTypes = {
+    /**
+     * The ID of this component, used to identify dash components
+     * in callbacks. The ID needs to be unique across all of the
+     * components in an app.
+     */
     id: PropTypes.string,
 
     /**
-     * An array of options
+     * An array of options {label: [string|number], value: [string|number]},
+     * an optional disabled field can be used for each option
      */
     options: PropTypes.arrayOf(
         PropTypes.exact({
@@ -152,6 +158,11 @@ Dropdown.propTypes = {
     ]),
 
     /**
+     * height of each option. Can be increased when label lengths would wrap around
+     */
+    optionHeight: PropTypes.number,
+
+    /**
      * className of the dropdown element
      */
     className: PropTypes.string,
@@ -188,6 +199,9 @@ Dropdown.propTypes = {
      */
     setProps: PropTypes.func,
 
+    /**
+     * Defines CSS styles which will override styles previously set.
+     */
     style: PropTypes.object,
 
     /**
@@ -214,4 +228,5 @@ Dropdown.defaultProps = {
     disabled: false,
     multi: false,
     searchable: true,
+    optionHeight: 35,
 };